+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
+## [26.3.9] - 2026-04-06
+
+### Added
+
+- Breadcrumbs
+
+### Changed
+
+- `Text` properties change handler
+- Anchor styling
+
+
+
 ## [26.3.8] - 2026-04-05
 
 ### Added